requirements.txt
PyQt5==5.15.11
websocket-client==1.8.0

